Updated Map & TP Count
Now if you click the map, it takes you into a battle. Have the same issue with having to re-navigate to the page to force the new image to render. Added tp numbers for total, and for battle checks. The progress bar isn't there, but it takes you out of battle once you got enough tp. Added lifetime tp total to complete tasks page